% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{cachemer}
\alias{cachemer}
\title{cachemer}
\format{\code{\link{R6Class}} object.}
\usage{
cachemer
}
\value{
Object of \code{\link{R6Class}} with methods for caching objects
}
\description{
TODO
}
\section{Fields}{

\describe{
\item{\code{new(path="config.yaml",overwrite=TRUE,env=new.env())}}{Initializes \code{\link{R6Class}} object}

\item{\code{summary}}{Prints summary of \code{\link{R6Class}} object}

\item{\code{getEnv}}{Gets environment shared between multiple `cachemer` objects}

\item{\code{share(nm,val)}}{Shares object between multiple `cachemer` objects}

\item{\code{getShared(nm)}}{Gets shared object}

\item{\code{cacheme(fun.name,fun.body,arguments,output=NULL,algo="md5")}}{Caches object/function}

\item{\code{lastCache}}{Gets last object added to cache}
}}

\examples{
\dontrun{

}

}
\keyword{data}
