% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{cachemerRef}
\alias{cachemerRef}
\title{ref}
\format{\code{\link{R6Class}} object.}
\usage{
cachemerRef
}
\value{
Object of \code{\link{R6Class}} with methods for caching objects
}
\description{
Inherits from \link[cachemeR]{cacher}
}
\keyword{data}
