% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isArgList.R
\name{evalOutput}
\alias{evalOutput}
\title{evalOutput}
\usage{
evalOutput(output, envir = parent.frame())
}
\arguments{
\item{output}{object of class `call`}

\item{envir}{environment}
}
\value{
value of `call` expression
}
\description{
Evaluate `call` expression even when some elements are listss
}
