% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhs_pipe.R
\name{getChainArgs}
\alias{getChainArgs}
\title{getChainArgs}
\usage{
getChainArgs(chain_parts)
}
\arguments{
\item{chain_parts}{\code{\link[cachemeR]{split_chain}} output}
}
\value{
list
}
\description{
Get arguments of function from \code{\link[cachemeR]{split_chain}} output
}
