% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_something.R
\name{is_parenthesized}
\alias{is_parenthesized}
\title{is_parenthesized}
\usage{
is_parenthesized(expr)
}
\arguments{
\item{expr}{}
}
\value{
logical - TRUE if expression is parenthesized, FALSE otherwise.
}
\description{
Determine whether an non-evaluated call is parenthesized. It is wrapper 
around unexported function \code{magrittr:::is_parenthesized}
}
