% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isArgList.R
\name{isListArg}
\alias{isListArg}
\title{isListArg}
\usage{
isListArg(x, envir)
}
\arguments{
\item{x}{obj of class `call` or `name``}

\item{envir}{environment in wchich}
}
\value{
logical
}
\description{
isListArg
}
