% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isArgList.R
\name{parseList}
\alias{parseList}
\title{parseList}
\usage{
parseList(x, envir)
}
\arguments{
\item{x}{object of class `call` or `name`}

\item{envir}{environment}
}
\value{
vector
}
\description{
parseList
}
