% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{cache}
\alias{cache}
\alias{\%c-\%}
\title{\%c-\%}
\usage{
cache(x, value)

x \%c-\% value
}
\arguments{
\item{x}{variable to assign output from `value`}

\item{value}{to be assigned to `x`}
}
\description{
cacheR pipe operator
}
