% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_restore.R
\name{saveCache}
\alias{saveCache}
\title{saveCache}
\usage{
saveCache(x, path, promises.env, sufix, prefix = "cachemer",
  plan = "multiprocess", workers = future::availableCores() - 1,
  logger.name = "test.logger", future.plan = "multiprocess")
}
\arguments{
\item{x}{}

\item{promises.env}{}

\item{plan}{}

\item{workers}{}

\item{logger.name}{}
}
\description{
Save object to rds
}
