% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhs_pipe.R
\name{split_chain}
\alias{split_chain}
\title{split_chain}
\usage{
split_chain(x, env)
}
\arguments{
\item{x}{}

\item{env}{}
}
\value{
list
}
\description{
Split \code{\link[base]{match.call}} into elements like 
\code{functions}, \code{pipes}, \code{values} or \code{others}
}
