% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFun.R
\name{testFun}
\alias{testFun}
\alias{testFun2}
\alias{doLm}
\alias{calculateFibonacci}
\title{testFun}
\usage{
testFun(a = 1:20, b = 0, c = list(d = 3, e = 5))

testFun2()

doLm(rows, cols)

calculateFibonacci(nterms)
}
\arguments{
\item{a}{numeric vector}

\item{b}{numeric value}

\item{c}{list}

\item{rows}{integer}

\item{cols}{integer}

\item{nterms}{integer}
}
\value{
numeric

model

vector
}
\description{
Test function
}
