% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFun.R
\name{testFun}
\alias{testFun}
\title{testFun}
\usage{
testFun(x = 1:20, b = 0, c = list(a = 2, b = 3))
}
\arguments{
\item{x}{vector}

\item{b}{numeric}

\item{c}{list}
}
\value{
numeric
}
\description{
Test function
}
\examples{
testFun()
}
