% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{cachemerRef}
\alias{cachemerRef}
\title{cachemerRef}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for caching objects
}
\description{
Inherits from \code{\link{cachemer}}
}
\keyword{data}
\section{Super class}{
\code{\link[cachemeR:cacher]{cachemeR::cacher}} -> \code{cachemerRef}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{cachemerRef$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="cachemeR" data-topic="cacher" data-id="cacheme">}\href{../../cachemeR/html/cacher.html#method-cacheme}{\code{cachemeR::cacher$cacheme()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cachemeR" data-topic="cacher" data-id="clear">}\href{../../cachemeR/html/cacher.html#method-clear}{\code{cachemeR::cacher$clear()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cachemeR" data-topic="cacher" data-id="initialize">}\href{../../cachemeR/html/cacher.html#method-initialize}{\code{cachemeR::cacher$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="cachemeR" data-topic="cacher" data-id="summary">}\href{../../cachemeR/html/cacher.html#method-summary}{\code{cachemeR::cacher$summary()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cachemerRef$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
