% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getArgNames.R
\name{getArgs}
\alias{getArgs}
\title{getArgs}
\usage{
getArgs(value, eval.calls = TRUE, env = parent.frame())
}
\arguments{
\item{value}{function call}

\item{eval.calls}{logical}

\item{env}{environment}
}
\value{
list
}
\description{
getArgs
}
\examples{
\dontrun{
testFun <- function(a = 1:20, b = list(c = 3, d = 5)) {
  x \%>\% sin \%>\% sum
}
getArgs(testFun())
}
}
