% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{\%c-\%}
\alias{\%c-\%}
\title{\%c-\%}
\usage{
x \%c-\% value
}
\arguments{
\item{x}{variable to assign output from `value`}

\item{value}{to be assigned to `x`}

\item{env}{environment}
}
\description{
cacheR pipe operator
}
