% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_restore.R
\name{restoreCache}
\alias{restoreCache}
\alias{getPattern}
\title{restoreCache}
\usage{
restoreCache(
  path,
  sufix = ".*",
  prefix = "cachemer",
  plan = "multisession",
  workers = future::availableCores() - 1,
  logger.name = "test.logger",
  future.plan = "multisession"
)

getPattern(prefix, sufix = ".*")
}
\arguments{
\item{path}{}

\item{sufix}{}

\item{prefix}{}

\item{plan}{}

\item{workers}{}

\item{logger.name}{}

\item{future.plan}{}
}
\value{
list

string
}
\description{
restoreCache

getPattern
}
