% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFun.R
\name{testFun}
\alias{testFun}
\alias{testFun2}
\title{testFun}
\usage{
testFun(a = 1:20, b = 0, c = list(d = 3, e = 5))

testFun2()
}
\arguments{
\item{b}{numeric}

\item{c}{list}

\item{x}{vector}
}
\value{
numeric
}
\description{
Test function
}
